requirements.txt
langchain
langchain-openai
langchain-community
pywebview 
